+++ packages/ubuntu_desktop_installer/test/identity/test_identity.mocks.dart
@@ -9,7 +9,7 @@ import 'dart:ui' as _i6;
 import 'package:mockito/mockito.dart' as _i1;
 import 'package:ubuntu_desktop_installer/pages/identity/identity_model.dart'
     as _i2;
-import 'package:ubuntu_desktop_installer/services.dart' as _i4;
+import 'package:ubuntu_provision/ubuntu_provision.dart' as _i4;
 import 'package:ubuntu_widgets/ubuntu_widgets.dart' as _i3;
 
 // ignore_for_file: type=lint
