+++ packages/ubuntu_desktop_installer/test/test_utils.mocks.dart
@@ -17,7 +17,6 @@ import 'package:ubuntu_desktop_installer/services/active_directory_service.dart'
     as _i8;
 import 'package:ubuntu_desktop_installer/services/config_service.dart' as _i10;
 import 'package:ubuntu_desktop_installer/services/desktop_service.dart' as _i11;
-import 'package:ubuntu_desktop_installer/services/identity_service.dart' as _i3;
 import 'package:ubuntu_desktop_installer/services/installer_service.dart'
     as _i15;
 import 'package:ubuntu_desktop_installer/services/journal_service.dart' as _i16;
@@ -26,6 +25,7 @@ import 'package:ubuntu_desktop_installer/services/product_service.dart' as _i6;
 import 'package:ubuntu_desktop_installer/services/session_service.dart' as _i22;
 import 'package:ubuntu_desktop_installer/services/storage_service.dart' as _i24;
 import 'package:ubuntu_desktop_installer/services/udev_service.dart' as _i7;
+import 'package:ubuntu_provision/src/services/identity_service.dart' as _i3;
 import 'package:ubuntu_provision/src/services/keyboard_service.dart' as _i17;
 import 'package:ubuntu_provision/src/services/locale_service.dart' as _i18;
 import 'package:ubuntu_provision/src/services/power_service.dart' as _i21;
