+++ packages/ubuntu_provision/lib/services.dart
@@ -1,3 +1,4 @@
+export 'src/services/identity_service.dart';
 export 'src/services/keyboard_service.dart';
 export 'src/services/locale_service.dart';
 export 'src/services/power_service.dart';
